split out sparse::table as a submodule of tensor::sparse
